+++ internal/dhtcrawler/ping.go
@@ -11,24 +11,6 @@ import (
 	"time"
 )
 
-func (c *crawler) reseedBootstrapNodes(ctx context.Context) {
-	interval := time.Duration(0)
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-time.After(interval):
-			for _, addr := range c.bootstrapNodes {
-				//if c.kTable.HasAddr(addr.Addr()) {
-				//	continue
-				//}
-				_ = c.peersForFindNode.In(ctx, peer{addr: addr})
-			}
-		}
-		interval = c.reseedBootstrapNodesInterval
-	}
-}
-
 func (c *crawler) awaitPeersForPing(ctx context.Context) {
 	args := dht.MsgArgs{
 		ID: c.kTable.Origin(),
