+++ internal/protocol/dht/ktable/btree/stats.go
@@ -12,8 +12,6 @@ type BucketStats struct {
 func (n *rootNode) Stats() Stats {
 	buckets := make([]BucketStats, 0, n.N())
 	for i := n.N() - 1; i >= 0; i-- {
-		//path := make(Bits, i+1)
-		//path[i] = Bit1
 		count := n.bucketCounts[i]
 		if count > 0 {
 			buckets = append(buckets, BucketStats{
